+++ sql-to-dbsp-compiler/SQL-compiler/src/main/java/org/dbsp/sqlCompiler/compiler/backend/rust/ToRustVisitor.java
@@ -208,7 +208,6 @@ public class ToRustVisitor extends CircuitVisitor {
 
     @Override
     public VisitDecision preorder(DBSPSinkOperator operator) {
-        this.writeComments(operator.query);
         this.writeComments(operator)
                 .append(operator.input().getName())
                 .append(".")