+++ sql-to-dbsp-compiler/SQL-compiler/src/test/java/org/dbsp/sqlCompiler/compiler/sql/SqlIoTest.java
@@ -470,6 +470,7 @@ public abstract class SqlIoTest extends BaseSQLTests {
                 if (line.contains("\t")) {
                     separator = "\t";
                     inHeader = false;
+                    continue;
                 }
             }
             if (line.contains("---")) {