+++ sql-to-dbsp-compiler/lib/sqllib/src/binary.rs
@@ -4,6 +4,8 @@ use dbsp_adapters::{DeserializeWithContext, SerializeWithContext, SqlSerdeConfig
 use serde::{Deserializer, Serializer};
 use size_of::SizeOf;
 use std::fmt::Debug;
+use hex::ToHex;
+use crate::some_function1;
 
 #[derive(
     Debug,
@@ -52,4 +54,41 @@ impl ByteArray {
     pub fn new(d: &[u8]) -> Self {
         Self { data: d.to_vec() }
     }
+
+    pub fn length(self: &Self) -> usize {
+        self.data.len()
+    }
+
+    pub fn zip<F>(self: &Self, other: &Self, op: F) -> ByteArray
+        where F: Fn(&u8, &u8) -> u8
+    {
+        let self_len = self.data.len();
+        let other_len = other.data.len();
+        if self_len != other_len {
+            panic!("Cannot operate on BINARY objects of different sizes {} and {}",
+                   self_len, other_len);
+        }
+        let result: Vec<u8> = self.data.iter().zip(other.data.iter()).map(|(l, r)| op(l, r)).collect();
+        ByteArray::new(&result)
+    }
+
+    pub fn and(self: &Self, other: &Self) -> Self {
+        self.zip(other, |left, right| left & right)
+    }
+
+    pub fn or(self: &Self, other: &Self) -> Self {
+        self.zip(other, |left, right| left | right)
+    }
+
+    pub fn xor(self: &Self, other: &Self) -> Self {
+        self.zip(other, |left, right| left ^ right)
+    }
 }
+
+pub fn to_hex_(value: ByteArray) -> String {
+    value.data.encode_hex::<String>()
+}
+
+some_function1!(to_hex, ByteArray, String);
+
+